requirements.txt
openai >= 1.0.0
pysolr
streamlit
PyMuPDF
nltk
rank_bm25
numpy
python-dotenv